change vacancies endpoints

+++ src/main/java/com/descenty/work_in_spring/company/service/CompanyService.java
@@ -1,5 +1,6 @@
 package com.descenty.work_in_spring.company.service;
 
+import com.descenty.work_in_spring.company.Company;
 import com.descenty.work_in_spring.company.dto.CompanyCreate;
 import com.descenty.work_in_spring.company.dto.CompanyDTO;
 import com.descenty.work_in_spring.area.repository.AreaRepository;
@@ -7,8 +8,11 @@ import com.descenty.work_in_spring.company.CompanyMapper;
 import com.descenty.work_in_spring.company.repository.CompanyRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,23 +32,28 @@ public class CompanyService {
     }
 
     @Transactional
-    public Optional<Long> create(Long areaId, CompanyCreate companyCreate) {
+    public Optional<Long> create(Long areaId, CompanyCreate companyCreate, UserDetails userDetails) {
         if (!areaRepository.existsById(areaId))
             return Optional.empty();
         return Optional.of(companyCreate).map(company -> {
             company.setAreaId(areaId);
+            company.setEmployersEmails(Collections.singletonList(userDetails.getUsername()));
             return companyMapper.toEntity(company);
-        }).map(companyRepository::save).map(company -> company.getId());
+        }).map(companyRepository::save).map(Company::getId);
     }
 
     @Transactional
-    public Optional<CompanyDTO> update(Long id, CompanyCreate companyCreate) {
+    public Optional<CompanyDTO> update(Long id, CompanyCreate companyCreate, UserDetails userDetails) {
+        if (!companyRepository.existsByIdAndEmployersEmailsContaining(id, userDetails.getUsername()))
+            return Optional.empty();
         return companyRepository.findById(id).map(company -> companyMapper.update(company, companyCreate))
                 .map(companyRepository::save).map(companyMapper::toDTO);
     }
 
     @Transactional
-    public Boolean delete(Long id) {
+    public boolean delete(Long id, UserDetails userDetails) {
+        if (!companyRepository.existsByIdAndEmployersEmailsContaining(id, userDetails.getUsername()))
+            return false;
         return companyRepository.findById(id).map(area -> {
             companyRepository.delete(area);
             return true;
